src/requirements.txt
APScheduler==3.10.*
SQLAlchemy==2.0.*
Flask==2.2.*
flask-session2==1.3.*
Flask-WTF==1.1.*
jsonschema==4.17.*
